third_party/WebKit/LayoutTests/fast/html/details-clone-expected.txt

Test to clone details and summary elements

PASS targetMarkerPseudoId is clonedMarkerPseudoId
PASS successfullyParsed is true

TEST COMPLETE
